Add Armor sorting
- Armor types are needed in more places now, so we move them to a
  different file in @/types
